+++ nuxeo/blob.py
@@ -4,13 +4,14 @@ from __future__ import unicode_literals
 import mimetypes
 import os
 import sys
+from io import StringIO
 
 try:
-    from StringIO import StringIO
+    from typing import Any, Dict, IO, Optional, Text, Union
 except ImportError:
-    from io import StringIO
+    pass
 
-__all__ = ('BatchBlob', 'Blob', 'BufferBlob', 'FileBlob')
+__all__ = ('BlobInfo', 'Blob', 'BufferBlob', 'FileBlob')
 
 
 WIN32_PATCHED_MIME_TYPES = {
@@ -28,6 +29,7 @@ WIN32_PATCHED_MIME_TYPES = {
 
 
 def guess_mime_type(filename):
+    # type: (Text) -> Text
     mime_type, _ = mimetypes.guess_type(filename)
     if mime_type:
         if sys.platform == 'win32':
@@ -46,47 +48,55 @@ class Blob(object):
     """
 
     def __init__(self):
+        # type: () -> None
         self._name = None
         self._size = 0
         self._data = ''
         self._mimetype = 'application/octet-stream'
 
     def get_data(self):
+        # type: () -> Union[IO[bytes], StringIO, Text]
         """ Data. """
         return self._data
 
     def get_mimetype(self):
+        # type: () -> Text
         """ Mimetype for the content. """
         return self._mimetype
 
     def get_name(self):
+        # type: () -> Text
         """ Return the name to use. """
         return self._name
 
     def get_size(self):
+        # type: () -> int
         """ Size in bytes. """
         return self._size
 
 
-class BatchBlob(Blob):
-    """ Representation of an already uploaded Blob. """
+class BlobInfo(object):
+    """ Contains info about an uploaded Blob. """
 
     def __init__(self, service, obj):
+        # type: (BatchUpload, Dict[Text, Any]) -> None
         self.service = service
         self.uploaded = obj.get('uploaded', 'true') == 'true'
         self.uploadType = obj['uploadType']
-        self._name = obj['name']
-        self._size = int(obj.get('size', 0))
-        self.uploadedSize = int(obj.get('uploadedSize', self._size))
+        self.name = obj['name']
+        self.size = int(obj.get('size', 0))
+        self.uploadedSize = int(obj.get('uploadedSize', self.size))
         self.fileIdx = int(obj['fileIdx'])
 
     def get_batch_id(self):
+        # type: () -> Text
         return self.service.get_batch_id()
 
     def to_json(self):
+        # type: () -> Dict[Text, Text]
         return {
             'upload-batch': self.get_batch_id(),
-            'upload-fileId': str(self.fileIdx),
+            'upload-fileId': str(self.fileIdx).decode('utf-8'),
         }
 
 
@@ -94,6 +104,7 @@ class BufferBlob(Blob):
     """ InMemory content to upload to Nuxeo. """
 
     def __init__(self, buf, name, mimetype='application/octet-stream'):
+        # type: (Text, Text, Text) -> None
         """
         :param buf: Content to upload to Nuxeo.
         :param name: Name to give to the file created on the server.
@@ -106,9 +117,11 @@ class BufferBlob(Blob):
         self._buffer = buf
 
     def get_data(self):
+        # type: () -> StringIO
         return StringIO(self._buffer)
 
     def get_size(self):
+        # type: () -> int
         return len(self._buffer)
 
 
@@ -116,6 +129,7 @@ class FileBlob(Blob):
     """ Represent a File as Blob for future upload. """
 
     def __init__(self, path, mimetype=None):
+        # type: (Text, Optional[Text]) -> None
         """
         :param path: To the file.
         :param mimetype: If not specified client will try to guess.
@@ -125,12 +139,10 @@ class FileBlob(Blob):
         self._path = path
         self._name = os.path.basename(self._path)
         self._size = os.path.getsize(self._path)
-        if mimetype is None:
-            self._mimetype = guess_mime_type(path)
-        else:
-            self._mimetype = mimetype
+        self._mimetype = mimetype or guess_mime_type(path)
 
     def get_data(self):
+        # type: () -> IO[bytes]
         """ Request data. """
 
         return open(self._path, 'rb')