+++ nuxeo/operation.py
@@ -1,19 +1,26 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
-from .blob import BatchBlob
+from .blob import BlobInfo
+
+try:
+    from typing import Any, Dict, List, Optional, Text, Union
+except ImportError:
+    pass
 
 __all__ = ('Operation',)
 
 
 class Operation(object):
     def __init__(self, nature, service):
+        # type: (Text, Nuxeo) -> None
         self.service = service
         self._type = nature
         self._params = dict()
         self._input = None
 
     def execute(self, **kwargs):
+        # type: (**Any) -> Union[Dict[Text, Any], Text]
         """
         Execute the operation on the server.
 
@@ -24,7 +31,7 @@ class Operation(object):
         url = None
         input_ = self._input
         params = self._params
-        if isinstance(input_, BatchBlob):
+        if isinstance(input_, BlobInfo):
             url = (self.service.rest_url
                    + 'upload/'
                    + input_.service.batchid
@@ -37,10 +44,12 @@ class Operation(object):
             self._type, url=url, params=params, op_input=input_, **kwargs)
 
     def input(self, input_):
+        # type: (Union[Text, List[Text], Dict[Text, Any]]) -> None
         """ Set the input for this operation. """
         self._input = input_
 
     def params(self, params=None):
+        # type: (Optional[Dict[Text, Any]]) -> Dict[Text, Any]
         """
         Get or add parameters to the operation.