+++ tests/conftest.py
@@ -1,12 +1,12 @@
 # coding: utf-8
-
 from __future__ import print_function
+from __future__ import unicode_literals
 
 import os
 import socket
-from requests import HTTPError
 
 import pytest
+from requests import HTTPError
 
 from nuxeo.blob import BufferBlob
 from nuxeo.document import Document